[frontend] Rename auth.tsx to client-auth.tsx
